requirements.txt
accelerate==1.10.0
asyncio==3.4.3
azure-cognitiveservices-speech==1.45.0
boto3==1.40.12
fastapi==0.119.1
google-auth==2.40.3
google-cloud-dialogflow-cx==1.42.0
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.29.0
gql==3.5.0
httpx==0.27.2
matplotlib==3.10.7
nemo-text-processing==1.1.0
number-parser==0.3.2
nvidia-riva-client==2.16.0
openai==1.55.1
pyannote.audio==3.3.2
redis==5.0.8
requests_toolbelt==1.0.0
torch==2.3.0
torchvision==0.18.0
transformers==4.40.2
uvicorn==0.38.0
websockets==12.0


